flask_web/requirement.txt
alembic==1.0.9
certifi==2019.3.9
cffi==1.12.3
Click==7.0
decorator==4.4.0
enum34==1.1.6
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.0
idna==2.8
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.9
MarkupSafe==1.1.1
pbr==5.1.3
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.3
sqlalchemy-migrate==0.12.0
sqlparse==0.3.0
Tempita==0.5.2
Werkzeug==0.15.2
